+++ riverpie_flutter/test/showcase_test.dart
@@ -96,13 +96,13 @@ void main() {
         notifier: counterNotifier,
         prev: 0,
         next: 2,
-        flagRebuild: [WidgetRebuildable<MyPage>()],
+        rebuild: [WidgetRebuildable<MyPage>()],
       ),
       ChangeEvent(
         notifier: counterNotifier,
         prev: 2,
         next: 4,
-        flagRebuild: [WidgetRebuildable<MyPage>()],
+        rebuild: [WidgetRebuildable<MyPage>()],
       ),
     ]);
   });